% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Test 2018/9/1
Loads a CSV file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to the CSV file (character)}
}
\value{
Returns a tibble from the CSV file
}
\description{
Reads and loads a CSV file defined by \code{filename} argument
Returns a tibble
If the path is incorrect the function will end with an error.
}
\examples{
\dontrun{
accident_2017 <- fars_read("accident_2017.csv.bz2")
}

}
